Reorganize tests
Reorganizes the structure of the tests modules to match the structure of
the ScrAPD ones. This helps newcomers understand the application
through the tests more easily.
